3961
Expand the steps in `update_html` and try to process each set of files
separately so we can see commonality.

The eventual goal is a script that can selectively process a subset of
files. But this is a good first step: I can at least easily comment out
different subsets.