3972 - drop ancient experiment for global variables
We never ended up using this, and all it has provided is a hole in the
type system.

Java's lesson is that globals are not essential. Any program that uses
globals can be rewritten to do without them. Just thread a world object
through the program, saving global state in it, and extracting relevant
bits of global state to pass into functions lower down the call chain.

Thanks Andrew Owen for the conversation that led me to focus on this
issue.